gsplat-rust-renderer: move tools/ CLI logic into apis/ thin shims
Follow-up to #50. These tools/ scripts kept their Tyro Config + main() + helpers
inline, so `beartype_this_package()` (dev-mode runtime type checking) never
instrumented them — the heavy jaxtyping annotations were effectively decorative.

Move the implementation of visualize_brush_training, log_splats_with_cameras,
calibration_scene, and relog_check into `gsplat_rust_renderer/apis/<name>.py` and
reduce each `tools/<name>.py` to a ~5-line Tyro entrypoint shim, matching the repo
convention (e.g. `simplecv/tools/view_exoego.py`).

Pure relocation, no logic change. Verified: ruff + pyrefly clean, 13 tests pass,
all three Tyro CLIs `--help`, and the apis modules import under beartype (so the
annotations are now enforced in dev). Addresses Copilot review comments #5-8 on #50.
