+++ packages/gsplat-rust-renderer/gsplat_rust_renderer/apis/__init__.py
@@ -1,0 +1,1 @@
+"""High-level Tyro Config + main() entrypoints (beartype-instrumented; tools/ are thin shims)."""
